+++ src/enemy/components.rs
@@ -3,22 +3,22 @@ use bevy::prelude::*;
 
 #[derive(Component)]
 pub struct Enemy {
-    pub health: f32,
-    pub speed: f32,
-    pub damage: f32,
-    pub xp: f32,
+    pub health: u32,
+    pub speed: u32,
+    pub damage: u32,
+    pub xp: u32,
     pub enemy_type: EnemyType,
 }
 
 #[derive(Component)]
 pub struct Trail {
-    pub damage: f32,
+    pub damage: u32,
     pub radius: f32,
 }
 
 #[derive(Component)]
 pub struct Collider {
-    pub radius: f32,
+    pub radius: u32,
 }
 
 #[derive(Component)]
